fix bug: closing fold w/ 'fold open' closed parent fold
When all the nested folds in a fold are open and 'fold open' is used,
only the fold you are on should close not it's parent fold.

Problem: parent fold was closing

Solution: use function that folded each nest fold recursively until the
fold was closed. Without folding the parent fold.